Adds user password change functionality (#37)
Introduces functionality allowing users to change their passwords.
It includes a new service for changing passwords after validating the old password,
a serializer for request data validation, and a view to expose the functionality through an API endpoint.
Adds tests to ensure the correct functionality of the new password change.